Solve exporting issues

+++ inc/minishell.h
@@ -6,7 +6,7 @@
 /*   By: xriera-c <xriera-c@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/26 16:58:52 by xriera-c          #+#    #+#             */
-/*   Updated: 2024/06/04 17:41:18 by xriera-c         ###   ########.fr       */
+/*   Updated: 2024/06/05 10:50:22 by xriera-c         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -213,7 +213,7 @@ char	*ft_getenv(const char *name, char **env);
 size_t	array_size(char **arr);
 int		new_path_arr(t_env *env_s, char *str);
 char	*get_name(char *str);
-size_t	find_equal_sign(char *str);
+int		find_equal_sign(char *str);
 
 //SIGNALS
